Alter_user_table + avatar table (hotfix)

+++ database/migrations/2020_01_16_151014_alter_users_table.php
@@ -14,7 +14,7 @@ class AlterUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('avatar_id');
+            $table->unsignedInteger('avatar_id');
             $table->foreign('avatar_id')->references('id')->on('avatares');
         });
     }
